# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Nourishment 2.0 API",
        "contact": {
            "email": "marekt94@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/mealsinday": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of all meals in day",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mealsinday"
                ],
                "summary": "List meals in day",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/meal.MealInDay"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update meal in day by JSON body (must contain ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mealsinday"
                ],
                "summary": "Update an existing meal in day",
                "parameters": [
                    {
                        "description": "MealInDay payload",
                        "name": "mealinday",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meal.MealInDay"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new meal in day from JSON body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mealsinday"
                ],
                "summary": "Create a new meal in day",
                "parameters": [
                    {
                        "description": "MealInDay payload",
                        "name": "mealinday",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meal.MealInDay"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/mealsinday/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single meal in day by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mealsinday"
                ],
                "summary": "Get meal in day by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MealInDay ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meal.MealInDay"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete meal in day by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mealsinday"
                ],
                "summary": "Delete a meal in day",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MealInDay ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of all products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "List products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/meal.Product"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update product by JSON body (must contain ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update an existing product",
                "parameters": [
                    {
                        "description": "Product payload",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meal.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product from JSON body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product payload",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meal.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single product by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meal.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete product by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "meal.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "meal.Meal": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "productsInMeal": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meal.ProductInMeal"
                    }
                },
                "recipe": {
                    "type": "string"
                }
            }
        },
        "meal.MealInDay": {
            "type": "object",
            "properties": {
                "afternoonSnack": {
                    "$ref": "#/definitions/meal.Meal"
                },
                "breakfast": {
                    "$ref": "#/definitions/meal.Meal"
                },
                "dinner": {
                    "$ref": "#/definitions/meal.Meal"
                },
                "factorAfternoonSnack": {
                    "type": "number",
                    "format": "float64"
                },
                "factorBreakfast": {
                    "type": "number",
                    "format": "float64"
                },
                "factorDinner": {
                    "type": "number",
                    "format": "float64"
                },
                "factorLunch": {
                    "type": "number",
                    "format": "float64"
                },
                "factorSecondBreakfast": {
                    "type": "number",
                    "format": "float64"
                },
                "factorSupper": {
                    "type": "number",
                    "format": "float64"
                },
                "for5Days": {
                    "description": "mapowane z CHAR(1) '1' lub '0'",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "lunch": {
                    "$ref": "#/definitions/meal.Meal"
                },
                "name": {
                    "type": "string"
                },
                "secondBreakfast": {
                    "$ref": "#/definitions/meal.Meal"
                },
                "supper": {
                    "$ref": "#/definitions/meal.Meal"
                }
            }
        },
        "meal.Product": {
            "type": "object",
            "properties": {
                "carbohydrates": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/meal.Category"
                },
                "fat": {
                    "type": "number"
                },
                "fiber": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "kcalPer100": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "proteins": {
                    "type": "number"
                },
                "salt": {
                    "type": "number"
                },
                "sugar": {
                    "type": "number"
                },
                "sugarAndCarb": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "meal.ProductInMeal": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/meal.Product"
                },
                "weight": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Get token from authorization request and place in \"Value\" field \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}